feat : get_diary contents
