+++ ai_model/app/__init__.py
@@ -4,8 +4,8 @@ from flask import Flask
 def create_app():
     app = Flask(__name__)
 
-    from .views import main_views
-    app.register_blueprint(main_views.bp)
+    from .views import text_views
+    app.register_blueprint(text_views.bp)
     # @app.route('/')
     # def hello():
     #     return 'Hello, My First Flask!'
