--- ai_model/app/views/main_views.py
@@ -1,13 +1,0 @@
-from flask import Blueprint
-
-bp = Blueprint('main', __name__, url_prefix='/')
-
-
-@bp.route('/hello')
-def hello_pybo():
-    return 'Hello, Pybo!'
-
-
-@bp.route('/')
-def hello_index():
-    return 'index'
